<chore> 패키지 구조 변경
* 패키지 구조가 혼잡하게 섞여 있어서 레이어 구조로 수정
* 프로젝트 특성상 서비스 단위가 작고 단순하기 때문에 추후 도메인이 확장될 가능성이 낮고 요청에 대한 흐름을 계층화하여 프로젝트 구조를 쉽게 파악하고 일관성을 높이기 위해 레이어 구조를 채택함
