--- src/main/java/com/wook/chap/dto/UrlDto.java
@@ -1,12 +1,0 @@
-package com.wook.chap.dto;
-
-import lombok.Data;
-
-import javax.validation.constraints.NotBlank;
-
-@Data
-public class UrlDto {
-
-    @NotBlank
-    private String requestURI;
-}
