--- src/main/java/com/wook/chap/entity/TimeEntity.java
@@ -1,28 +1,0 @@
-package com.wook.chap.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@MappedSuperclass
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-public class TimeEntity {
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createTime;
-
-    @LastModifiedDate
-    private LocalDateTime lastModifiedTime;
-}
