--- src/main/java/com/wook/chap/entity/Url.java
@@ -1,36 +1,0 @@
-package com.wook.chap.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name="url")
-public class Url extends TimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="url_id")
-    private Long id;
-
-    @Column(name="shorts_url", unique = true)
-    private String shortURL;
-
-    @Column(name="original_url",length = 2048)
-    private String originalURL;
-
-    @Column(name="click_count")
-    private int clickCount;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-}
