--- src/main/java/com/wook/chap/enums/Authority.java
@@ -1,6 +1,0 @@
-package com.wook.chap.enums;
-
-public enum Authority {
-
-    USER, MANAGER
-}
