--- src/main/java/com/wook/chap/jwt/JwtAccessDeniedHandler.java
@@ -1,24 +1,0 @@
-package com.wook.chap.jwt;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-@Slf4j
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-    @Override
-    public void handle(HttpServletRequest request,
-                       HttpServletResponse response,
-                       AccessDeniedException accessDeniedException) throws IOException, ServletException {
-
-        log.debug("권한 오류 발생 {}", accessDeniedException.getMessage());
-        response.sendError(HttpServletResponse.SC_FORBIDDEN);
-    }
-}
