--- src/main/java/com/wook/chap/jwt/JwtAuthenticationEntryPoint.java
@@ -1,34 +1,0 @@
-package com.wook.chap.jwt;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-import static com.wook.chap.jwt.JwtFilter.AUTHORIZATION_HEADER;
-
-@Component
-@Slf4j
-public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-
-    @Override
-    public void commence(HttpServletRequest request,
-                         HttpServletResponse response,
-                         AuthenticationException authException) throws IOException, ServletException {
-
-        log.debug("인증 오류 발생 {}",authException.getMessage());
-        Cookie cookie = new Cookie(AUTHORIZATION_HEADER, null);
-        cookie.setMaxAge(0);
-        cookie.setPath("/");
-
-        response.addCookie(cookie);
-        response.sendRedirect("/login/authenticate");
-    }
-}
