+++ src/main/java/com/wook/chap/model/dto/UrlDto.java
@@ -1,0 +1,12 @@
+package com.wook.chap.model.dto;
+
+import lombok.Data;
+
+import javax.validation.constraints.NotBlank;
+
+@Data
+public class UrlDto {
+
+    @NotBlank
+    private String requestURI;
+}
