+++ src/main/java/com/wook/chap/model/entity/Member.java
@@ -1,0 +1,38 @@
+package com.wook.chap.model.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Member extends TimeEntity{
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="member_id")
+    private Long id;
+
+    @Column(unique = true)
+    private String name;
+
+    private String email;
+
+    private String password;
+
+    private String nickname;
+
+
+    public Member(String name, String email, String password, String nickname) {
+        this.name = name;
+        this.email = email;
+        this.password = password;
+        this.nickname = nickname;
+    }
+}
