+++ src/main/java/com/wook/chap/model/enums/Authority.java
@@ -1,0 +1,6 @@
+package com.wook.chap.model.enums;
+
+public enum Authority {
+
+    USER, MANAGER
+}
