+++ src/main/java/com/wook/chap/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package com.wook.chap.repository;
+
+import com.wook.chap.model.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+
+    Optional<Member> findByName(String memberName);
+}
