--- src/main/java/com/wook/chap/web/controller/HomeController.java
@@ -1,19 +1,0 @@
-package com.wook.chap.web.controller;
-
-import com.wook.chap.dto.UrlDto;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String home(Model model) {
-        UrlDto urlDto = new UrlDto();
-        model.addAttribute("urlDto", urlDto);
-        return "registerUrl";
-    }
-
-
-}
