+++ src/main/resources/errors.properties
@@ -2,21 +2,21 @@ invalid.user=해당 아이디가 존재하지 않습니다.
 invalid.password=비밀번호가 틀렸습니다.
 invalid.url=형식에 맞지 않는 URL 입니다.
 
-NotBlank.loginForm.membername=아이디에 빈칸을 입력하셨습니다.
-NotBlank.loginForm.password=비밀번호에 빈칸을 입력하셨습니다.
+NotBlank.loginDto.membername=아이디에 빈칸을 입력하셨습니다.
+NotBlank.loginDto.password=비밀번호에 빈칸을 입력하셨습니다.
 
 member.duplicate=이미 가입된 아이디가 존재합니다.
 
 password.wrongexpression=비밀번호는 영문+특수문자+숫자 8자로 구성되어야 합니다.
 password.notmatch=비밀번호가 서로 일치하지 않습니다.
 
-NotBlank.signUpForm.membername=아이디에 빈칸을 입력하셨습니다.
-NotBlank.signUpForm.password=비밀번호에 빈칸을 입력하셨습니다.
-NotBlank.signUpForm.repeatPassword=비밀번호 재입력란에 빈칸을 입력하셨습니다.
-NotBlank.signUpForm.email=이메일에 빈칸을 입력하셨습니다.
-NotBlank.signUpForm.nickname=닉네임에 빈칸을 입력하셨습니다.
+NotBlank.signUpDto.membername=아이디에 빈칸을 입력하셨습니다.
+NotBlank.signUpDto.password=비밀번호에 빈칸을 입력하셨습니다.
+NotBlank.signUpDto.repeatPassword=비밀번호 재입력란에 빈칸을 입력하셨습니다.
+NotBlank.signUpDto.email=이메일에 빈칸을 입력하셨습니다.
+NotBlank.signUpDto.nickname=닉네임에 빈칸을 입력하셨습니다.
 NotBlank.urlDto.requestURI=URL에 빈칸을 입력하셨습니다.
-Email.signUpForm.email=이메일 형식이 아닙니다.
+Email.signUpDto.email=이메일 형식이 아닙니다.
 
-Size.signUpForm.membername=아이디는 8~20 글자를 입력해주세요.
-Size.signUpForm.nickname=닉네임은 5~10 글자를 입력해주세요.
+Size.signUpDto.membername=아이디는 8~20 글자를 입력해주세요.
+Size.signUpDto.nickname=닉네임은 5~10 글자를 입력해주세요.
