+++ src/main/resources/templates/form/loginForm.html
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
 
     <link rel="stylesheet" href="../../static/css/login/style.css" th:href="@{/css/login/style.css}">
-    <link rel="stylesheet" href="../../static/css/login/loginForm.css" th:href="@{/css/login/loginForm.css}">
+    <link rel="stylesheet" href="../../static/css/login/loginDto.css" th:href="@{/css/login/loginDto.css}">
     <link rel="icon" href="data:,">
 
 </head>
@@ -31,7 +31,7 @@
             <div class="col-md-6 col-lg-4">
                 <div class="login-wrap p-0">
                     <h3 class="mb-4 text-center">회원이신가요?</h3>
-                    <form action="#" th:action th:object="${loginForm}" class="signin-form" method="post">
+                    <form action="#" th:action th:object="${loginDto}" class="signin-form" method="post">
 
                         <div th:if="${#fields.hasGlobalErrors()}">
                             <p class="field-error-global" th:each="err :${#fields.globalErrors()}" th:text="${err}">글로벌 오류 메시지</p>
