+++ src/test/java/com/wook/chap/web/utils/UrlInspectorTest.java
@@ -1,5 +1,6 @@
 package com.wook.chap.web.utils;
 
+import com.wook.chap.utils.UrlInspector;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
